Adding tests for RecipeApi
